mock 関数を関連サブモジュール全体で利用できるように場所を移動 #3
